Updates #13696
Changed map function to create a tmp file with just the first 5 rows of the CSV to parse for mapping use.

Increased the memory limit and script time out.

Working up to ~6500 records on a shared LAMP server. I think it's hitting the HTTP timeout of 5 min rather than the PHP one at that point.

Working up to 20k records on MAMP/local.

git-svn-id: https://central.tri.be/svn/premium-plugins/code/events-importer/trunk@1147 75e66736-2afa-4fbe-8bd5-dce959cd71f9