show no-results notice on ajax load of day view, fix skipped post
updates #19067
When the day view is loaded via ajax, tribe_is_day() was returning false, so
the no-results notices for day views didn't display. Filtered said function
to return true.

The ajax handler for the day view has a mysterious the_post(), causing the
view to skip the first event on the page. Added a rewind so that the
first post will also display.