Work to fix #23260 (title of most recent event post showing for events views in some themes).
This commit doesn't actually change anything as the "trigger" has been commented out (line 77) - committing it only in case we want to resurrect this sort of approach in the future with some refinements. Right now, filtering the_title is too much of a shotgun approach and is likely to cause new problems even if it solves the original one.