remove the extra caching layer of events options
The plugin was fetching the tribe_events_calendar_options and storing it in a static variable.
This didn't give any performance boost, since the data is already cached in memory by WordPress.
But, it made the plugin incompatible with switch_to_blog() calls, since the extra cache layer
did not update when blogs were switched. This could result in data corruption if attempting
to set options while switched to another blog.