Solving the Map Problem without an AJAX request As Sam pointed out on the Slack chat, doing an extra request was bad performance wise (My bad).
So I reverted the AJAX request and found out how to solve this issue without any extra requests.
The solution is on the month.php template class, which was using `$wp_query->query` that at some point lost the `meta_query`, so I just appended this variable there and BOOM now it's working.

Updates #31027