Assigns the google map object back to tribeEventsSingleMap global JS var
This allows plugin or theme code to modify the map. The tribeEventsSingleMap var will have all the data needed: initial zoom level, addresses, and the google map object itself.

For example, we may want to disable the scrollwheel. With this commit, the following can be called from outside The Event Calendar:
`tribeEventsSingleMap.map.setOptions({scrollwheel:false});`