added check for empty event object - 34113
The bug at [Notices generated when looking at week view](https://central.tri.be/issues/34113) seems to be raised when using the Week View. That is hooking into the `tribe_events_event_classes` filter and trying to compute on a null event object. Essentially there seemed not to be a bail check in place to avoid trying to manipulate classes for events not in the week view.