Localize the events url base and use that to generate event URLs
On sites where there aren't any events, pagination links aren't displayed which in turn breaks the `base_url` generation. To prevent that from happening, let's localize the base event URL and start with _that_ rather than `/`.  This ensures that the event URL is properly formed even in the absence of events.