Performance enhancements that target month view
**The Cost_Utils class**

`get_cost_by_func` was pretty slow for two reasons: `array_merge` is kind of slow _and_ when there are a large number of events, fetching every non-unique cost means we have to loop over and parse each one.  That's expensive. Let's select distinct costs and avoid doing `array_merge`.

**Month view class**

When fetching the daily events, we were looping over every event in the month and determining if it fell on the given day. Instead, we could reduce the number of times we loop by _instead_ looping over all the events in the month and breaking them out into the days that they fall on. We get the same data, it is just built in reverse and requires fewer loops.

See: https://github.com/moderntribe/the-events-calendar/pull/192 and https://central.tri.be/issues/32067