Prevent throwing function declaration fatals when running event-tickets
In an effort to prevent fatals when running event-tickets and a pre-4.0 version of the-events-calendar, we need to wrap a number of functions with if statements. While this won't help any TEC install less than 3.12.4, we can at least prevent the issue with the latest version.

See: https://central.tri.be/issues/40986