Don't change the visibility of these methods. Make new methods
To more easily test `get_updates()` and `constant_updates()`, those methods were converted to public. This causes fatals when Pro 3.12.x is active because Pro has a class that extends `Tribe__Events__Updater` and overrides the `protected` methods.

The fix is to retain those methods as protected and make new ones that are public.

See: https://central.tri.be/issues/40281