Don't hide the default address fields if a default venue is selected
Just because the site owner wants a default venue to be selected when creating an event _doesn't_ mean that they don't want default address information when creating a new event in the dashboard.  Let's keep the address fields visible so defaults can be selected.

See: https://central.tri.be/issues/29366#note-17