Remove date filters on import pages now that query is cleaned up
With [this fix](https://github.com/moderntribe/the-events-calendar/pull/445), we don't really need filters on the import pages. In fact, excluding them is necessary as previous false successes were caused by the date of the event and the date injected into the query working through crazy random happenstance. When looking for importable objects, we won't need to have a hijacked query.

See: https://central.tri.be/issues/41385