Treat raw_filters like filters when generating the view_more link
When specifying filters via the Events List shortcode, those filters are collected under the `raw_filters` index rather than the json_encoded `filters` index in the `$instance` variable used when rendering the layout.

This change continues to use `filters` if it is set, but then falls back to using `raw_filters` if that is set.

See: https://central.tri.be/issues/37529