Multi-day All-day events may require an adjusted date by eod cutoff
All-day Multi-day events have their end dates displayed individually in a few locations - when viewing a single event and viewing the event in the admin dashboard. In those instances, the true end date is not what is desired, but, rather an end date that observes the end of day cutoff.

This changeset adds a new method ( `tribe_get_display_end_date()` ) that calls `tribe_get_end_date()` and adusts the output based on the multi-day/eod status of the event whose end date is being fetched.

See: https://central.tri.be/issues/40555#note-16