#84202 Undo Datepicker Format Code Change from M17.16
_Ref:_ [C#84202](http://central.tri.be/issues/84202)

This commit may seem weird—but it is indeed all that’s required to undo
the support for “MM.DD.YYYY”-style date formats (i.e., formats
containing a dot instead of the usual dash or slash) in the Month View
Tribe Bar.

It thus also fixes the bug in #86937.