Fix bug where the should_hide_recurrence() / tribeHideRecurrence in some circumstances
The should_hide_recurrence() / tribeHideRecurrence functionality breaks when this widget is shown on the same page as the Month event calendar.

I have the list widget shown in the sidebar of my site (showing the 3 latest upcoming events in a particular category) with the "Recurring event instances” setting at Events => Settings => General _checked_ so it would hide recurring instances after the first upcoming. This works as intended on all pages except for the main events page that's set to show the monthly calendar view, because it appears that setting is completely ignored on that specific page.

This one change in the code has it trigger the logic for following that setting as expected as before it appears the Monthly view was overwriting that setting (so it shows all events) and then the widget wasn't reporting itself as a widget so it was left using what previous setting.

I would love to see this bug fix included in a future version as this fixed a problem on a site that I have in production. Thanks!