Make sure to set is_home if we are on the front page
In order to prevent any 404 error code we need to set `is_home` to `true` if.
we are on the front page, which means the main reading settings are set
to display a page that is displaying an events page, this should turn:

- `is_home`
- `is_front_page`

To `true`.

See: https://central.tri.be/issues/90001