s/EventStartDate/_EventStartDate/ in magic post property
When building the query used to lookup adjacent links, it appears
that an empty string is always used for the `_EventStartDate`
value and unexpected events are used in the previous/next links
on single event pages.

When `$post_obj->EventStartDate` is called, post meta is checked
for a meta key that does not exist. Adjusting this to the correct
meta key, `$post_obj->_EventStartDate` provides a correct date
and results in properly generated links.