Dynamic determine the number of events
Between now and a month ahead with a 5 day interval between each element
of the array.

At the end instead of compare with a fixed number we compare with the
filter result as the number of elements is dynamic due the months on the
yars does not have the same number of days.