Change behavior related to what is considered Past on Events listing
There was a problem with the old mplementation in general, that has been with us for a while.

On the AJAX request, we use "Now" as the cut-off from the events and on the browser refresh, we use the beginning of the day which creates the inconsistencies on some days and depending on the time you are looking.

_Ref:_ [C#123950](http://central.tri.be/issues/123950)