docs(filter documentation): Add `null` as second possible type for the `$query` parameter
In the `tribe_events_query_event_utc_start_date` filter.
When used in the context of the `Tribe__Events__Query::pre_get_posts` method the `$query` parameter
will be set to the currently filtered query. When used in other contexts, e.g. the Month view, there
will be no query to filter and passing the global one is misleading as to what is being affected.

re https://central.tri.be/issues/125196