fix(iCal feed): Fix iCal feed generation to use ORM arguments correctly.
Switching to the ORM the main query arguments, the contents of the `query` property, and the main
query variables, the contents of the `query_vars` property, are not always in sync. While building
the ORM query a number of variables are set on the query in the `query_vars` property that are not
set on the `query` argument. Since the iCal feed function was reading the `query` property and not
the `query_vars` one, some arguments that produced the correct event list in the view template would
not make it into the `query` property resulting in a different set of events. The fix simply changes
the iCal feed generation to use the `query_vars` property. Furthermore a small "fix" to the
arguments is needed when producing the iCal feed for the List view to remove the `eventDate`
argument, used in the Day view.

re https://central.tri.be/issues/125433