fix(ORM): Initialize the date meta keys used by the ORM from Timezone Settings.
In its first implementation the ORM would always use UTC dates to fetch events. This would correctly
align with the Timezone Setting "use site-wide timezone" but it would invalidate the usage of the
"use manual timezones for each event" setting. This commit adds an initialization of the meta keys
the ORM should use depending on the setting value and modifies the doc-blocks and accessory code
that was relying on that setting to adapt. I've also added a `use_utc( bool $use_utc )` method to
the Event ORM to allow forcing that behavior on a per-instance basis.

re https://central.tri.be/issues/125196