fix(List View): Ensure the List view AJAX handler is using the same criteria as the main query.
There is disparity between the filters applied to the List view when rendered via PHP initial state
(main query) and when rendered in the context of an AJAX request. This commit fixed that behavior by
making sure the AJAX handling method uses the same filtering criteria as the main query.

re https://central.tri.be/issues/123950