feat(template-tags/date.php): Added the `tribe_event_ends_on` function.
The `tribe_event_ends_on` function will assert whether an event ends on a date or not. The check is
timezone, and timezone settings, aware. As an example a 4/10 2pm to 7pm, Los Angeles timezone, event
does not end on 4/11 if using the Los Angeles timezone, it will end on 4/11 (at 4am) if using the
Paris timezone. What timezone will be used for the comparison depends on the site timezone and the
site timezone settings.

re https://central.tri.be/issues/125587