fix(List View): Add more robust List view hashing system.
When no bar date is provided the List view would default to use "now" and pivot to define current
and past events. Since that "now" is down to the second paginating back and forth in the List view,
supported by AJAX methods, would generate different hashes on each request thus resending the user
back to page 1 on each request. The List view uses an hash to spot if anyrelevat filter parameter
has changed, since that hash would include the date a date of "now" would always generate a new
hash. Rounding the date, even when defaulting to "now", down to the closest 30' makes it so that the
hashing will be consistent, time-wise, for the pretty reasonable time of 30 minutes. If the user is
still browsing back and forth in the following 30' the user will be simply taken back to page 1 to
be shown the udpated list version. What we had before was, furthermore, ignoring Filter Bar filters
in the hashing.

re https://central.tri.be/issues/123950