refactor(Views v2): move 1st degree filters into service provider, remove slug registration
This refactor tries to make the most of PHP 5.6 support by removing slug-based registration, to use
the class-based registration instead, to leverage di52 automagical dependency resolution.
Additionally filters and actions previously registered in the classes are moved in the service
provider to model the service provider as the main wiring component for any first-level
filter/action used by Views v2.