fix(Repositories\Event): filter meta input during creation before checking it
This fix makes the event creation, via the Repository, a bit more fluid by filtering the meta input
that will be used to create the event before cheking it. The case this fix will allow to cover is a
pretty common one where one might want to create an event by using the
following syntax:
```php
$event_id = tribe_events()->set_args([
	'title' => 'Test Event',
	'start_date' => 'tomorrow 9am',
	'duration' => 7200 // 2 hours,
])->create();
```

In the previous version this syntax would have failed to create an event
as the existence of an end date was checked **before** it could be built
from the `start_date` and `duration` information.