test(Views v2): Filter the `home_url` in the base test case to make sure printed URLs are consistent
The home URL will change depending on the setup running it (e.g. CI or local); this fix makes it so
that the URLs will be consistent, if built using the `home_url` function, across envs defaulting the
root `home_url`, in the context of View tests, to the `http://test.tri.be` one.