Avoid fiddling with the start date of query if is a query into the past
The start_date query param was being overridden on the list view for all queries that didn't have the start date set. However, on the base list view, a query into the past is needed to determine if the previous link is necessary, so overriding the start_date breaks it. This change prevents the override from occuring on queries that are looking into the past.

Additionally, this PR adds an acceptance test suite with tests to verify that the previous and next links are present at the right moments on initial page load.

See: https://central.tri.be/issues/127477