fix(rewrites): fix category parsing issue
The commit udpates the code and the tests to make sure we're producing canonical URls, and
processing them, the same way WordPress does it by trying to find a match top to bottom. As a
consequence I've updated some of the expectations to make make sure they match what the actual
canonical URLs would be. I've also added a check on the category to make sure we're rewriting only
if they exist.