fix(Tribe__Events__Rewrite.php): readd request cache in `get_bases`.
In release/B19.07.1 we've removed the caching, on a per-request base with a simple array, done in
the `get_bases` method. Getting the bases is a costly operation (there are 3 filters applying on it)
that should be done as little as possible. Furthermore the caching was removed as consequence of an
eager call to the `setup` method in the `Tribe__Rewrite::instance` method that is removed in the
linked PR (https://github.com/moderntribe/tribe-common/pull/1043), thus this caching can be
reinstated.