test(views_integration): add post remapping support for snapshot testing
Snapshot testing, due to its own nature, is broke when dealing with any kind of data dynamic in
nature that will vary from run to run. Events, created in the tests, will pose this challenge. This
commit tries to get the best of both worlds by providing a was to remap real posts to fixed payloads
and a way to test the dynamic content of a View.