fix(EA Imports): re-run async process support check when stopping impotrs
This commit forces a new check for async processing support when the user clicks the link to stop
and clear current asynchronous import processes (in Events > Settings > Imports). This is done under
the assumption that, if the user is clicking the link in the first place, somethiing is not working
as intended. Since the feature detection is cached this forced re-check might immediately solve an
issue where there is a discrepancy between the import system and its support. Even if the user has
selected the "Asynchronous" import processing system, if the feature detection fails, the system
will actually use cron (this pre-dates this fix, it's part of the fallback mechanism of the
processing system).