fix(shortcode) allow disabling URL push
Ticket: n/a

This commit updates the Javascript code for the calendar and the
localized `tribe_js_config` object, to allo fine-grained disabling of
the URL update when navigating a view produced by the [tribe_events]
shortcode.

By default nothing changes, but this allows developers to override the
settings by filtering the contents of the `tribe_js_config` object by
filtering on the `tribve_events_js_config` filter.