fix(timezone notice): The timezone notice responds incorrectly.
The timezone notice appears to respond incorrectly when on an event edit page (admin). The function checks the event timezone properly, but when the notice is set, it uses the site timezone. This is all well and good until we're using "manual timezones for each event" - in which case the site timezone is fine but the event one is not (defaults to UTC+0).

Then we get messages like "...and avoid using a UTC timezone offset such as “America/Los_Angeles”. This just confuses folks.

Proposed fix: store the timezone string when we check it and reuse it in the notice.