fix(Collections) add and use Lazy_Post_Collection
This commit adds, and uses, the `Lazy_Post_Collection` class to
correctly handle serialization and unserialization of decorated post
objects.
Calls to `serialize` on `WP_Post` objects decorated by methods like
`tribe_get_event` or `tribe_get_venue_object` will not throw any error
and will produce a serialized version of the post object w/ all the
added properties added to it.

During unserialization, though, the additional, decorating, properties
will not be recognized by the `unserialize` function and the object
instance will fail to build, returning `false` as result of the
unserialization.

This commit adds extra care when dealing with post objects to store only
an aray of post IDs and a callback that should be used to hydrate them
during unserialization.