test(wpunit) add test coverage for tribe_get_region
Add a test to cover the fix in PR
https://github.com/moderntribe/the-events-calendar/pull/3015.
When the Venue `_VenueStateProvince` meta is not set, and the Venue
state is not the US, then the Venue region would be pulled for the
global Venue post object.

This commit does not fix the issue, fixed in the original PR, but add
tests for it to avoid regressions.