fix(Views v2) filter the View query arg slug
Issue: https://moderntribe.atlassian.net/browse/BTRIA-271

This commit allows filtering the query argument we use to detect a View
request in the URL.

Some other plugins might use `view` for their own purposes and our
usages, and detection of View requests, might be offset or wrong.

The commit adds the filtering function and upates the code where it was
using the query arg directly.

The changes are small as 99% of Views v2 code does not look up the query
arg directly and relies, instead, on the Context.