Code that fixes the incorrect data being passed. And code that checks for site data if event data is missing.
The attr() function was being passed an extra param, so it was returning incorrect default values.
The timezone info is saved to the block as attributes - so it's not saved to meta and it's not saved if the block isn't changed. This means that if we get a default value back, we have to check for it in Event->Settings->General.

[TEC-2964]