fix(ignored-posts): Filter wp_count_posts() to include a dummy count for tribe-ignored status to prevent notices.
By making the status available in `show_in_admin_status_list` we tell WP to han dle it for every post table - but WP doesn't check to ensure it's actually there.
So it tries to access a non-existent property and throws a notice. This hooks into `wp_count_posts()` and adds a dummy (0) value for it to prevent notices.