Correct the today link on shortcodes with category(s).
While we corrected the prev/next links in [ECP-492] ( see https://github.com/moderntribe/the-events-calendar/pull/3304 ). The issue still remains for the "today" button.
This implements a similar fix for the today button.

I looked for additional instances of this url handling on my local and didn't find any. If anyone knows of another it would make sense to add it to this PR.