fix(I18n.php) correctly backup locale
Before version 4.7 WordPress would use nested arrays to keep track of
filters.
The code fixed in this commit pre-dates 4.7 and took a somewhat hybrid
approach to backing up the filter setting it to an empty array if not
set.

In previous versions of WordPress a "compatibilty layer" would convert
array-format filters to the new, object-based, implementation. The
removal of that compatibility layer and our code unknowingly reliance on
the fact that there would always be something hooked on the `locale`
filter did protect us from issues so fare.

This commit updates the logic that was setting the backup for the
`locale` filter to an empty array if nothing was hooked to it, to an
empty hook object.