feat(Views/By_Day_View) alter the iCal export args
This commit modifies the By Day View, parent to the Month and Week Views
currently, to make sure the repository args will be set to fetch the
events between the View grid start and end date.

By-day Views, are the only Views that do not work like a list and do
require this.