refactor(iCalendar/Request) build the View as the manager would
This commit changes the code that is responsible for detecting, and
handling an iCalendar export request, by leveraging the fact that any
iCalendar GET request will fully provide the View URL providing, thus,
all the arguments the View requires to set itself up.

With some little modifications to the repository arguments, implemented
in previous commits, the View will be in the best position to return the
matching event post IDs.