fix(V2/View.php) correctly restore the loop vars
This commit updates the `View::setup_the_loop` and
`View::restore_the_loop` code to make sure the backup and restoring of
the variables we need to alter in the context of the View loop and then
restore after the loop completed will be correctly restored.

The correct reset of the `$_SERVER['REQUEST_URI]` variable, in
particular, should solve a number of similar issues.