fix(Views/V2/TemplateBootstrap) archive body class
This commit fixes an issue where our code would remove the `archive`
body class if the request is not for a taxonomy page.

To detect whether a request is for a taxonomy page or not, we would use
the `is_tax` template tag that is, though, not the correct way to do it.

The `is_tax` template tag will return `true` if:
1. The current request is for a **custom** taxonomy term page.

Since the post `tag` and `category` terms are **not** part of a custom
taxonomy, the call to `is_tax` would return `false` and we would
erroneously remove the `archive` body class in that case.