fix(Views/Day_View) correctly handle start/end of day
This commit wraps and adds tests for the community PR
https://github.com/the-events-calendar/the-events-calendar/pull/3410
to add tests and changelog entry.

While a bit complex the issue can be reproduced as follows:
1. Have "now" be 10pm EST time.
2. Have the server timezone be UTC
3. In UTC timezone, 10pm EST today will be 2am tomorrow.
4. When fed to the `tribe_begginning_of_day` and `tribe_end_of_day`
   functions, then the start and end dates would be built on tomorrow.

The behaviour is incorrect and fixed by the original PR by making sure
that, when no explicit date is provided to the Day View, the "now" is
before localized to the site timezone.