fix(Traits/HTML_Cache) ensure HTML key consistency
This fix solves an issue where the HTML cache key generated for a view
Context at the start of the request would **never** match the one
generated at the end of the request for the same Context. The way the
HTML cache is built, anything that **adds** locations to the Context or
the Context locaction cache will yield a different cache key. This
commit fixes it by caching, in non-persistent cache, the HTML Cache Key
using the context SPL object hash as key.

The context is an **input** to the View, what we want to make sure is
that we do not re-calculate the same HTML provided the same **input**.
The input is the Context in its initial state, not the final one.