fix(CT1/Modeld) ensure string dates
In both the Event and the Occurrence models, depending how and where
they are built, the date attributes (`start_date`, `end_date` and their
UTC versions) are hinted, in the `@property` blocks as `string` but
might, in reality, be `DateTimeInterface` instances.

This fix makes sure the hinted property type is honored by means of a
trait that will run that date to string conversion when the date
attributes are, actually, accessed.