fix(Models/Occurrence) align to latest data
This fix solves 2 issues in the current implementation.
The first is that the `get_occurrences` method would `align_event_meta`
after generating the new, and up-to-date, Occurrences, but **before**
those are actually saved to the db. The Event would, then, be aligned on
the **current**, thus out-of-date Occurrence state creating a number of
issues down the road and ignoring, in fact, any update to the Event and
its Occurrences after the creation.

The second one is a fix that will actually call `align_event_meta`
passing it the first and last Occurrence references, coming from the
still-to-commit values, to run the updates.