refactor(CT1) alt. update flow first draft
This commit adds the base code that will provide the alternate update
flow in the context of TEC and CT1.

As a proof-of-concept, this code is not the final version yet, but
contains pretty much all the required ideas and concepts.

Just a note: enable the alt. update flow with this entry in the
`wp-config.php` file:

```
define( 'TEC_CUSTOM_TABLES_V1_ALT_UPDATE_FLOW', true );
```