refactor(CT1/Models/Occurrence) search match
When searching for an existing Occurence matching the data of the
insertion that should be created, do use the `hash`, but use the dates
and post ID.

This change would impact the execution speed of the search, since the
`hash` column is indexed, but would not add more queries.

This way of fetching a matching Occurrence is more robust and flexible
than the hash-based one, and will result in more Occurrences being
recycled (thus less disruption).