refactor(CT1/Models/Occurrences) search by dates
Change the logic used by the `get_occurrences` method to find existing
Occurrences that might match the one that _should_ be inserted to drop
the use of `hash` and rely on the use of dates.

This makes the check more robust as the information we care about are
post ID and dates, and simpler to manage as transfering Occurrences will
require very simple queries (in place of a complex set of Occurrence
Model building and such).