refactor(CT1/Models/Occurrence) filter Occ. match
Change the logic that will find an existing Occurrence in the database
when trying to insert a new one to make is to TEC, that only handles
single Events, will keep re-using the same Occurrence over and over.

Filter the logic to allow other plugins, like ECP to change the matching
rules.