fix(CT1/Repository) re-connect
The previous work had disconnected the repostitory API, the
`tribe_events()` function from the RBE code and would not have
immediate results due to the upsert functions running at the end of the
request.

This commit fixes this and re-connnects the Custom Tables v1 code to the
`tribe_events()` API to make the effect instant.