fix(CT1/Updates) known range calculation
When Events are inserted or updated in the database, then two options
are saved to keep track of the earliest start date and the latest end
date; these will be used in Views, e.g. Month view, to know whether we
should allow the user to paginate forward or not.
The default logic would use the Events meta, from the `postmeta` table
to calculate those, and this logic would fail in the context of RBE,
where the post meta is present only for the first Event, and not for
each one of its Occurrences.

This commit adds a filter into the default logic to use Occurrences
information to update the values.

Furthermore, this commit fixes an issue where the values would not be
updated when using the Blocks Editor.