feat(CT1/Models/Builder) add where_raw method
While the (Query) `Builder` and its API get us a long way, they might,
at times, not be enough for some more fine-grained, one-of-a-kind
queries.
This commit adds a the `Builder::where_raw` method to allow writing
particular WHERE clauses, in the context of the Builder that would allow
a complex API.
Under the hood, the method will use the `wpdb::prepare()` one and will
just conveniently insert the prepared query in the context of a query
being built.